Add Vulkan headless and offscreen rendering

+++ src/mbgl/renderer/layers/collision_layer_tweaker.cpp
@@ -69,12 +69,12 @@ void CollisionLayerTweaker::execute(LayerGroupBase& layerGroup, const PaintParam
 
 #ifdef MLN_RENDER_BACKEND_VULKAN
         const auto& globalUniforms = context.getGlobalUniformBuffers();
-        for (int i = 0; i < globalUniforms.allocatedSize(); ++i) {
+        for (size_t i = 0; i < globalUniforms.allocatedSize(); ++i) {
             if (globalUniforms.get(i)) drawableUniforms.set(i, globalUniforms.get(i));
         }
 
         const auto& layerUniforms = layerGroup.mutableUniformBuffers();
-        for (int i = 0; i < layerUniforms.allocatedSize(); ++i) {
+        for (size_t i = 0; i < layerUniforms.allocatedSize(); ++i) {
             if (layerUniforms.get(i)) drawableUniforms.set(i, layerUniforms.get(i));
         }
 #endif
