+++ src/mbgl/renderer/layers/heatmap_texture_layer_tweaker.cpp
@@ -46,11 +46,11 @@ void HeatmapTextureLayerTweaker::execute(LayerGroupBase& layerGroup, const Paint
         auto& drawableUniforms = drawable.mutableUniformBuffers();
 
         const auto& globalUniforms = parameters.context.getGlobalUniformBuffers();
-        for (int i = 0; i < globalUniforms.allocatedSize(); ++i) {
+        for (size_t i = 0; i < globalUniforms.allocatedSize(); ++i) {
             if (globalUniforms.get(i)) drawableUniforms.set(i, globalUniforms.get(i));
         }
 
-        for (int i = 0; i < layerUniforms.allocatedSize(); ++i) {
+        for (size_t i = 0; i < layerUniforms.allocatedSize(); ++i) {
             if (layerUniforms.get(i)) drawableUniforms.set(i, layerUniforms.get(i));
         }
     });
