+++ src/mbgl/renderer/layers/hillshade_layer_tweaker.cpp
@@ -71,11 +71,11 @@ void HillshadeLayerTweaker::execute(LayerGroupBase& layerGroup, const PaintParam
 
 #ifdef MLN_RENDER_BACKEND_VULKAN
         const auto& globalUniforms = parameters.context.getGlobalUniformBuffers();
-        for (int i = 0; i < globalUniforms.allocatedSize(); ++i) {
+        for (size_t i = 0; i < globalUniforms.allocatedSize(); ++i) {
             if (globalUniforms.get(i)) drawableUniforms.set(i, globalUniforms.get(i));
         }
 
-        for (int i = 0; i < layerUniforms.allocatedSize(); ++i) {
+        for (size_t i = 0; i < layerUniforms.allocatedSize(); ++i) {
             if (layerUniforms.get(i)) drawableUniforms.set(i, layerUniforms.get(i));
         }
 #endif
