+++ src/mbgl/renderer/layers/hillshade_prepare_layer_tweaker.cpp
@@ -65,11 +65,11 @@ void HillshadePrepareLayerTweaker::execute(LayerGroupBase& layerGroup, const Pai
         auto& layerUniforms = layerGroup.mutableUniformBuffers();
 
         const auto& globalUniforms = parameters.context.getGlobalUniformBuffers();
-        for (int i = 0; i < globalUniforms.allocatedSize(); ++i) {
+        for (size_t i = 0; i < globalUniforms.allocatedSize(); ++i) {
             if (globalUniforms.get(i)) drawableUniforms.set(i, globalUniforms.get(i));
         }
 
-        for (int i = 0; i < layerUniforms.allocatedSize(); ++i) {
+        for (size_t i = 0; i < layerUniforms.allocatedSize(); ++i) {
             if (layerUniforms.get(i)) drawableUniforms.set(i, layerUniforms.get(i));
         }
 #endif
