+++ src/mbgl/renderer/layers/symbol_layer_tweaker.cpp
@@ -137,11 +137,11 @@ void SymbolLayerTweaker::execute(LayerGroupBase& layerGroup, const PaintParamete
         
 #ifdef MLN_RENDER_BACKEND_VULKAN
         const auto& globalUniforms = context.getGlobalUniformBuffers();
-        for (int i = 0; i < globalUniforms.allocatedSize(); ++i) {
+        for (size_t i = 0; i < globalUniforms.allocatedSize(); ++i) {
             if (globalUniforms.get(i)) drawableUniforms.set(i, globalUniforms.get(i));
         }
 
-        for (int i = 0; i < layerUniforms.allocatedSize(); ++i) {
+        for (size_t i = 0; i < layerUniforms.allocatedSize(); ++i) {
             if (layerUniforms.get(i)) drawableUniforms.set(i, layerUniforms.get(i));
         }
 #endif
